Implemented value switch for 'a' and 'b'
This commit introduces code to switch the values of variables 'a' and 'b' using tuple unpacking. The completion adheres to the requirement of a 5-line program and provides the expected result of swapping the values of 'a' and 'b'. The commit message concisely describes the purpose of the code addition.