Implemented common_elements function for finding common elements
This commit adds the common_elements function to '3-common_elements.py.' The function efficiently calculates the common elements between two sets using the set intersection operation (&). The resulting set containing the common elements is returned as the output of the function. The commit message concisely describes the purpose and approach of the code addition.