Implemented number_keys function for counting dictionary keys
This commit introduces the number_keys function to '5-number_keys.py.' The function computes the number of keys present in the provided dictionary by utilizing the len() function on the dictionary object. The count of keys is returned as the output of the function. The commit message concisely highlights the purpose and method of the code addition.